// employee/src/request-workflow-type/common/ts/enum/request-workflow-type.enum.ts
export enum RequestWorkFlowTypeEnum {
  LEAVE_REQUEST = 'LEAVE_REQUEST',
  MISSION_REQUEST = 'MISSION_REQUEST',
  OVERTIME_REQUEST = 'OVERTIME_REQUEST',
  BORROW = 'BORROW/PAYBACK_HOUR',
  PAYBACK_HOUR = 'PAYBACK_HOUR',
  MISSED_SCAN = 'MISSED_SCAN',
  RESIGNATION_REQUEST = 'RESIGNATION_REQUEST',
  MOVEMENT = 'MOVEMENT',
  SALARY_ADJUSTMENT = 'SALARY_ADJUSTMENT',
  EMPLOYEE_INFO_UPDATE = 'EMPLOYEE_INFO_UPDATE',
  WARNING = 'WARNING',
  DAY_OFF_REQUEST = 'DAY_OFF_REQUEST',
  PAYROLL_BENEFIT_ADJUSTMENT = 'PAYROLL_BENEFIT_ADJUSTMENT',
  PAYROLL_BENEFIT = 'PAYROLL_BENEFIT',
  PAYROLL_DEDUCTION = 'PAYROLL_DEDUCTION'
}