// employee/src/leave/mission-request/enum/mission-request-duration-type.enum.ts
export enum MissionRequestDurationTypeEnEnum {
  DATE_RANGE = 'DATE_RANGE',
  FIRST_HALF_DAY = 'FIRST_HALF_DAY',
  SECOND_HALF_DAY = 'SECOND_HALF_DAY',
  TIME = 'TIME'
}

export enum MissionRequestDurationTypeKhEnum {
  DATE_RANGE = 'ចន្លេាះពេល',
  FIRST_HALF_DAY = 'ពាក់កណ្តាលថ្ងៃទីមួយ',
  SECOND_HALF_DAY = 'ពាក់កណ្តាលថ្ងៃទីពីរ',
  TIME = 'ម៉ោង'
}